Extract per-server secret helper (oh-tab-4ar)
